+++ tests/dataflow_test/test_nn_dataflow.py
@@ -60,16 +60,7 @@ class TestNNDataflow(unittest.TestCase):
                          noc_hop=0,
                          unit_static=0)
 
-        self.options = Option(sw_gbuf_bypass=(False, False, False),
-                              sw_solve_loopblocking=False,
-                              hw_access_forwarding=False,
-                              hw_gbuf_sharing=False,
-                              partition_hybrid=False,
-                              partition_batch=False,
-                              partition_ifmaps=False,
-                              ntops=1,
-                              nprocesses=1,
-                              verbose=False)
+        self.options = Option()
 
     def test_invalid_network(self):
         ''' Invalid network argument. '''
@@ -108,13 +99,6 @@ class TestNNDataflow(unittest.TestCase):
 
         options = Option(sw_gbuf_bypass=(True, True, True),
                          sw_solve_loopblocking=True,
-                         hw_access_forwarding=False,
-                         hw_gbuf_sharing=False,
-                         partition_hybrid=False,
-                         partition_batch=False,
-                         partition_ifmaps=False,
-                         ntops=1,
-                         nprocesses=1,
                          verbose=True)
 
         nnd = NNDataflow(network, batch_size, self.resource, self.cost,
@@ -391,14 +375,7 @@ class TestNNDataflow(unittest.TestCase):
 
         options = Option(sw_gbuf_bypass=(True, True, True),
                          sw_solve_loopblocking=True,
-                         hw_access_forwarding=False,
-                         hw_gbuf_sharing=False,
-                         partition_hybrid=True,
-                         partition_batch=False,
-                         partition_ifmaps=False,
-                         ntops=1,
-                         nprocesses=1,
-                         verbose=False)
+                         partition_hybrid=True)
 
         nnd = NNDataflow(network, batch_size, resource, cost,
                          self.map_strategy)