+++ CHANGELOG.md
@@ -12,7 +12,7 @@ List of major changes and improvements
   - Buffer sharing scheme.
 
 
-## [v1.4 -- v1.5] -- 2017-08-01
+## [v1.4 -- v1.5] -- 2017-08-05
 
 ### Added
 
@@ -43,6 +43,8 @@ List of major changes and improvements
 
   - Add negative operation to `PhyDim2`.
 
+  - Add default arguments to `Option`.
+
 - Test.
 
   - Add unit tests.
@@ -63,6 +65,8 @@ List of major changes and improvements
 
   - Add `type` attribute to `NodeRegion` to differentiate processing and data
     node regions in `Resource`.
+
+  - Change default cost of the NoC hop traversal.
 
 - Software models:
 
@@ -98,6 +102,9 @@ List of major changes and improvements
   - Check ifmap layout against layer parameters in `Scheduling`.
 
   - Add number of nodes to scheduling result.
+
+  - Add `type` attribute to `DataLayout` to denote the type of the reside
+    region.
 
 - Software engineering.